feat: put gps functions into seperate file
