+++ ESP32/esp32_cyclone/include/sensors/gps.h
@@ -3,9 +3,6 @@
 
 #include "TinyGPS++.h"
 
-// GPS Serial
-TinyGPSPlus gps;
-
 struct gpsData
 {
   double latitude;
@@ -13,12 +10,11 @@ struct gpsData
   double height;
   time_t time;
 };
-gpsData gpsdata;
 
 // setup GPS serial connection
 void setupGPS();
-void readGPSData();
-void displayInfo();
-void updateAllData();
+void readGPSData(TinyGPSPlus &gps, gpsData &gpsdata);
+void displayInfo(TinyGPSPlus &gps, gpsData &gpsdata);
+void updateAllData(TinyGPSPlus &gps, gpsData &gpsdata);
 
 #endif
